several pies on the plot
