+++ src/plugins/d3/renderer/hooks/useSeries/prepareSeries.ts
@@ -1,7 +1,6 @@
 import cloneDeep from 'lodash/cloneDeep';
 import get from 'lodash/get';
 import type {ScaleOrdinal} from 'd3';
-import {scaleOrdinal} from 'd3';
 
 import type {
     BarXSeries,
@@ -11,16 +10,14 @@ import type {
     LineSeries,
     PieSeries,
 } from '../../../../../types';
-import {getRandomCKId} from '../../../../../utils';
 
-import {DEFAULT_PALETTE} from '../../constants';
-import type {PreparedLegend, PreparedPieSeries, PreparedSeries} from './types';
+import type {PreparedLegend, PreparedSeries} from './types';
 import {prepareLineSeries} from './prepare-line-series';
 import {prepareBarXSeries} from './prepare-bar-x';
 import {prepareBarYSeries} from './prepare-bar-y';
 import {prepareLegendSymbol} from './utils';
 import {ChartKitError} from '../../../../../libs';
-import {DEFAULT_DATALABELS_PADDING, DEFAULT_DATALABELS_STYLE} from './constants';
+import {preparePieSeries} from './prepare-pie';
 
 type PrepareAxisRelatedSeriesArgs = {
     colorScale: ScaleOrdinal<string, string>;
@@ -41,56 +38,6 @@ function prepareAxisRelatedSeries(args: PrepareAxisRelatedSeriesArgs): PreparedS
     };
 
     return [preparedSeries];
-}
-
-type PreparePieSeriesArgs = {
-    series: PieSeries;
-    legend: PreparedLegend;
-};
-
-function preparePieSeries(args: PreparePieSeriesArgs) {
-    const {series, legend} = args;
-    const dataNames = series.data.map((d) => d.name);
-    const colorScale = scaleOrdinal(dataNames, DEFAULT_PALETTE);
-    const stackId = getRandomCKId();
-
-    const preparedSeries: PreparedSeries[] = series.data.map<PreparedPieSeries>((dataItem) => {
-        const result: PreparedPieSeries = {
-            type: 'pie',
-            data: dataItem,
-            dataLabels: {
-                enabled: get(series, 'dataLabels.enabled', true),
-                style: Object.assign({}, DEFAULT_DATALABELS_STYLE, series.dataLabels?.style),
-                padding: get(series, 'dataLabels.padding', DEFAULT_DATALABELS_PADDING),
-                allowOverlap: get(series, 'dataLabels.allowOverlap', false),
-                connectorPadding: get(series, 'dataLabels.connectorPadding', 5),
-                connectorShape: get(series, 'dataLabels.connectorShape', 'polyline'),
-                distance: get(series, 'dataLabels.distance', 30),
-                softConnector: get(series, 'dataLabels.softConnector', true),
-            },
-            label: dataItem.label,
-            value: dataItem.value,
-            visible: typeof dataItem.visible === 'boolean' ? dataItem.visible : true,
-            name: dataItem.name,
-            id: '',
-            color: dataItem.color || colorScale(dataItem.name),
-            legend: {
-                enabled: get(series, 'legend.enabled', legend.enabled),
-                symbol: prepareLegendSymbol(series),
-            },
-            center: series.center || ['50%', '50%'],
-            borderColor: series.borderColor || '',
-            borderRadius: series.borderRadius ?? 0,
-            borderWidth: series.borderWidth ?? 1,
-            radius: series.radius || '100%',
-            innerRadius: series.innerRadius || 0,
-            stackId,
-        };
-
-        return result;
-    });
-
-    return preparedSeries;
 }
 
 export function prepareSeries(args: {
